Move DEBUG code paths From the main SceneDelegate to a new DebuggingSceneDelegate subclass to separate debug- and test-specific code from production code
