+++ EssentialApp/EssentialApp/SceneDelegate.swift
@@ -9,6 +9,7 @@ import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     var window: UIWindow?
+    var localStoreURL: URL = NSPersistentContainer.defaultDirectoryURL().appendingPathComponent("feed-store.sqlite")
 
     func scene(_ scene: UIScene, willConnectTo _: UISceneSession, options _: UIScene.ConnectionOptions) {
         guard let _ = (scene as? UIWindowScene) else { return }
@@ -20,14 +21,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         let remoteClient = makeRemoteClient()
         let remoteFeedLoader = RemoteFeedLoader(url: remoteURL, client: remoteClient)
         let remoteImageLoader = RemoteFeedImageDataLoader(client: remoteClient)
-
-        let localStoreURL = NSPersistentContainer.defaultDirectoryURL().appendingPathComponent("feed-store.sqlite")
-
-        #if DEBUG
-            if CommandLine.arguments.contains("-reset") {
-                try? FileManager.default.removeItem(at: localStoreURL)
-            }
-        #endif
 
         let localStore = try! CoreDataFeedStore(storeURL: localStoreURL)
         let localFeedLoader = LocalFeedLoader(store: localStore, currentDate: Date.init)
@@ -45,25 +38,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         )
     }
 
-    private func makeRemoteClient() -> HTTPClient {
-        #if DEBUG
-            if UserDefaults.standard.string(forKey: "connectivity") == "offline" {
-                return AlwaysFailingHTTPClient()
-            }
-        #endif
+    func makeRemoteClient() -> HTTPClient {
         return URLSessionHTTPClient(session: URLSession(configuration: .ephemeral))
     }
 }
-
-#if DEBUG
-    private class AlwaysFailingHTTPClient: HTTPClient {
-        private class Task: HTTPClientTask {
-            func cancel() {}
-        }
-
-        func get(from _: URL, completion: @escaping (Result<(Data, HTTPURLResponse), Error>) -> Void) -> HTTPClientTask {
-            completion(.failure(NSError(domain: "com.alfredohdz.essential-app.offline", code: 1)))
-            return Task()
-        }
-    }
-#endif
